+++ CHANGELOG.md
@@ -7,8 +7,10 @@ Scalyr Agent 2 Changes By Release
 Packaged by Arthur Kamalov <arthur@scalyr.com> on Aug 30, 2021 21:00 -0800
 --->
 
-Bug fix:
+Other:
+* Update Windows Event Log monitor to emit a warning if ``maxmimum_records_per_source`` config option is set to a non-default value when using a new event API where that config option has no affect.
 
+Bug fixes:
 * Don't log "skipping copying log lines" messages in case number of last produced bytes is 0.
 
 ## 2.1.21 "Ultramarine" - Jun 1, 2021